% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flag}
\alias{flag}
\title{True Positive and False Positive for SYP}
\usage{
flag(data, sample, TP, FP, gene_name, ratio, TP_flag)
}
\arguments{
\item{data}{The dataset that will be used. Ensure that the columns are named appropriately (gene_name for the proteins, and SYP_ratio for the ratio of interest)}

\item{sample}{The name of the sample (e.g. SYP or NES)}

\item{TP}{The list of true positive proteins. Ensure that the column name is gene_name.}

\item{FP}{The list of false positive proteins. Ensure that the column name is gene_name.}

\item{ratio}{The column that contains the ratio of interest (SYP or NES)}

\item{TP_flag}{The name of the column for the true positive flag (TP_SYP or TP_NES), so whether this protein is a TP for SYP or NES respectively}
}
\value{
The initial dataset, including TPR and FPR
}
\description{
This function calculates the True Positive Rate (TPR) and False Positive Rate (FPR) for the SYP dataset
}
