% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{opt_cutoff}
\alias{opt_cutoff}
\title{Graph to calculate the optimal Cutoff}
\usage{
opt_cutoff(
  data,
  sample,
  col1 = "#002C34",
  col2 = "#4F0433",
  title = "Optimal cutoff"
)
}
\arguments{
\item{data}{The dataset that will be used. Ensure that the columns are named appropriately (FPR and TPR).}

\item{sample}{The name of the sample (e.g. SYP or NES)}

\item{col1}{The color of the points in the graph (optional).}

\item{col2}{The color of the vertical line in the graph (optional).}

\item{title}{The title of the graph (optional).}
}
\value{
A ggplot2 object showing the graph to calculate the optimal cutoff.
}
\description{
This function creates a graph to calculate the optimal cutoff for the dataset. The optimal cutoff is the point on the ROC curve that maximises the TPR-FPR value. The function calculates the log2 of the normalised ratio and the TPR-FPR value, and plots the graph.
}
