% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{roc}
\alias{roc}
\title{ROC Curve}
\usage{
roc(
  data,
  FPR = "FPR",
  TPR = "TPR",
  title = "ROC Curve",
  col1 = "#002C34",
  col2 = "#4F0433"
)
}
\arguments{
\item{data}{The dataset that will be used. Ensure that the columns are named appropriately (FPR and TPR).}

\item{FPR}{The name of the column in data that contains the False Positive Rate (optional if named appropriately).}

\item{TPR}{The name of the column in data that contains the True Positive Rate (optional if named appropriately).}

\item{title}{The title of the ROC curve (optional).}

\item{col1}{The color of the points in the ROC curve (optional).}

\item{col2}{The color of the diagonal line in the ROC curve (optional).}
}
\value{
A ggplot2 object showing the ROC curve.
}
\description{
This function creates a ROC curve using the FPR and TPR columns in the dataset.
}
