% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_documentation.R
\docType{package}
\name{TurboIDanalysis}
\alias{TurboIDanalysis-package}
\alias{TurboIDanalysis}
\title{TurboID Analysis Package}
\description{
This package provides the functions needed to assign TurboID hits to either a
true or false positive category, and to the create the ROC curves described
in Cho et al., 2020 (Nat Protoc).
}
